Remove started default parameters (#915)
* Remove started default parameters

* Update the stop VM condition

* Update test case

+++ plugins/modules/azure_rm_virtualmachine.py
@@ -55,7 +55,6 @@ options:
             - Whether the VM is started or stopped.
             - Set to (true) with I(state=present) to start the VM.
             - Set to C(false) to stop the VM.
-        default: true
         type: bool
     allocated:
         description:
@@ -897,7 +896,7 @@ class AzureRMVirtualMachine(AzureRMModuleBase):
             subnet_name=dict(type='str', aliases=['subnet']),
             allocated=dict(type='bool', default=True),
             restarted=dict(type='bool', default=False),
-            started=dict(type='bool', default=True),
+            started=dict(type='bool'),
             generalized=dict(type='bool', default=False),
             data_disks=dict(type='list'),
             plan=dict(type='dict'),
@@ -1219,7 +1218,7 @@ class AzureRMVirtualMachine(AzureRMModuleBase):
                              .format(self.name, vm_dict['powerstate']))
                     changed = True
                     powerstate_change = 'deallocated'
-                elif not self.started and vm_dict['powerstate'] == 'running':
+                elif self.started is not None and not self.started and vm_dict['powerstate'] == 'running':
                     self.log("CHANGED: virtual machine {0} running and requested state 'stopped'".format(self.name))
                     changed = True
                     powerstate_change = 'poweroff'
